requirements.txt
fastapi
uvicorn
requests
python-dotenv
neo4j
networkx
openai
